% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_create_image_builder_streaming_url}
\alias{appstream_create_image_builder_streaming_url}
\title{Creates a URL to start an image builder streaming session}
\usage{
appstream_create_image_builder_streaming_url(Name, Validity)
}
\arguments{
\item{Name}{[required] The name of the image builder.}

\item{Validity}{The time that the streaming URL will be valid, in seconds. Specify a
value between 1 and 604800 seconds. The default is 3600 seconds.}
}
\description{
Creates a URL to start an image builder streaming session.
}
\section{Request syntax}{
\preformatted{svc$create_image_builder_streaming_url(
  Name = "string",
  Validity = 123
)
}
}

\keyword{internal}
