% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_describe_fleets}
\alias{appstream_describe_fleets}
\title{Retrieves a list that describes one or more specified fleets, if the
fleet names are provided}
\usage{
appstream_describe_fleets(Names, NextToken)
}
\arguments{
\item{Names}{The names of the fleets to describe.}

\item{NextToken}{The pagination token to use to retrieve the next page of results for
this operation. If this value is null, it retrieves the first page.}
}
\description{
Retrieves a list that describes one or more specified fleets, if the
fleet names are provided. Otherwise, all fleets in the account are
described.
}
\section{Request syntax}{
\preformatted{appstream$describe_fleets(
  Names = list(
    "string"
  ),
  NextToken = "string"
)
}
}

\keyword{internal}
