% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_start_image_builder}
\alias{appstream_start_image_builder}
\title{Starts the specified image builder}
\usage{
appstream_start_image_builder(Name, AppstreamAgentVersion)
}
\arguments{
\item{Name}{[required] The name of the image builder.}

\item{AppstreamAgentVersion}{The version of the AppStream 2.0 agent to use for this image builder. To
use the latest version of the AppStream 2.0 agent, specify [LATEST].}
}
\description{
Starts the specified image builder.
}
\section{Request syntax}{
\preformatted{appstream$start_image_builder(
  Name = "string",
  AppstreamAgentVersion = "string"
)
}
}

\keyword{internal}
