% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appstream_operations.R
\name{appstream_update_image_permissions}
\alias{appstream_update_image_permissions}
\title{Adds or updates permissions for the specified private image}
\usage{
appstream_update_image_permissions(Name, SharedAccountId,
  ImagePermissions)
}
\arguments{
\item{Name}{[required] The name of the private image.}

\item{SharedAccountId}{[required] The 12-digit identifier of the AWS account for which you want add or
update image permissions.}

\item{ImagePermissions}{[required] The permissions for the image.}
}
\description{
Adds or updates permissions for the specified private image.
}
\section{Request syntax}{
\preformatted{appstream$update_image_permissions(
  Name = "string",
  SharedAccountId = "string",
  ImagePermissions = list(
    allowFleet = TRUE|FALSE,
    allowImageBuilder = TRUE|FALSE
  )
)
}
}

\keyword{internal}
