% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_delete_folder}
\alias{workdocs_delete_folder}
\title{Permanently deletes the specified folder and its contents}
\usage{
workdocs_delete_folder(AuthenticationToken, FolderId)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}

\item{FolderId}{[required] The ID of the folder.}
}
\description{
Permanently deletes the specified folder and its contents.
}
\section{Request syntax}{
\preformatted{workdocs$delete_folder(
  AuthenticationToken = "string",
  FolderId = "string"
)
}
}

\keyword{internal}
