% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workdocs_operations.R
\name{workdocs_update_document}
\alias{workdocs_update_document}
\title{Updates the specified attributes of a document}
\usage{
workdocs_update_document(AuthenticationToken, DocumentId, Name,
  ParentFolderId, ResourceState)
}
\arguments{
\item{AuthenticationToken}{Amazon WorkDocs authentication token. Do not set this field when using
administrative API actions, as in accessing the API using AWS
credentials.}

\item{DocumentId}{[required] The ID of the document.}

\item{Name}{The name of the document.}

\item{ParentFolderId}{The ID of the parent folder.}

\item{ResourceState}{The resource state of the document. Only ACTIVE and RECYCLED are
supported.}
}
\description{
Updates the specified attributes of a document. The user must have
access to both the document and its parent folder, if applicable.
}
\section{Request syntax}{
\preformatted{workdocs$update_document(
  AuthenticationToken = "string",
  DocumentId = "string",
  Name = "string",
  ParentFolderId = "string",
  ResourceState = "ACTIVE"|"RESTORING"|"RECYCLING"|"RECYCLED"
)
}
}

\keyword{internal}
