% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspaces_operations.R
\name{workspaces_describe_tags}
\alias{workspaces_describe_tags}
\title{Describes the specified tags for the specified WorkSpace}
\usage{
workspaces_describe_tags(ResourceId)
}
\arguments{
\item{ResourceId}{[required] The identifier of the WorkSpace. To find this ID, use
DescribeWorkspaces.}
}
\description{
Describes the specified tags for the specified WorkSpace.
}
\section{Request syntax}{
\preformatted{svc$describe_tags(
  ResourceId = "string"
)
}
}

\keyword{internal}
